refactor(shelf): use ShelfId and rename outbound repo impl to adapter
- Replace Shelf Long id with ShelfId value object and update repository contract

- Update JPA-backed repository implementation to accept ShelfId and unwrap for persistence lookups

- Rename infrastructure package from port.outbound to adapter.outbound

- Fix docs wording: facade described as inbound adapter (not inbound port)
