+++ src/main/java/com/penrose/bibby/library/stacks/shelf/core/domain/valueobject/ShelfId.java
@@ -1,0 +1,4 @@
+package com.penrose.bibby.library.stacks.shelf.core.domain.valueobject;
+
+public record ShelfId(Long shelfId) {
+}
